validate category and product

+++ src/main/java/com/bookshop/controllers/ProductController.java
@@ -1,9 +1,17 @@
 package com.bookshop.controllers;
 
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Optional;
-
+import com.bookshop.dao.Category;
+import com.bookshop.dao.Product;
+import com.bookshop.dao.ProductImage;
+import com.bookshop.dto.ProductDTO;
+import com.bookshop.dto.ProductDetail;
+import com.bookshop.exceptions.DuplicateRecordException;
+import com.bookshop.exceptions.InvalidException;
+import com.bookshop.exceptions.NotFoundException;
+import com.bookshop.helpers.CheckValid;
+import com.bookshop.helpers.ConvertObject;
+import com.bookshop.repositories.CategoryRepository;
+import com.bookshop.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -11,26 +19,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
-import com.bookshop.dao.Category;
-import com.bookshop.dao.Product;
-import com.bookshop.dao.ProductImage;
-import com.bookshop.dto.ProductDTO;
-import com.bookshop.dto.ProductDetail;
-import com.bookshop.exceptions.DuplicateRecordException;
-import com.bookshop.exceptions.NotFoundException;
-import com.bookshop.helpers.ConvertObject;
-import com.bookshop.repositories.CategoryRepository;
-import com.bookshop.repositories.ProductRepository;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/products")
@@ -44,7 +37,27 @@ public class ProductController {
     private CategoryRepository categoryRepository;
 
     @GetMapping
-    public ResponseEntity<?> getAllProducts(@RequestParam(name = "page", required = false) Integer pageNum) {
+    public ResponseEntity<?> getAllProducts(@RequestParam(name = "page", required = false) Integer pageNum,
+                                            @RequestParam(name = "type", required = false) String type,
+                                            @RequestParam(name = "search", required = false) String search) {
+        if (search != null) {
+            List<Product> products = productRepository.findByTitleContaining(search);
+
+            if (products.size() == 0) {
+                return ResponseEntity.noContent().build();
+            }
+            return ResponseEntity.ok().body(products);
+        }
+        if (type != null) {
+            if (type.compareTo("without-image") == 0) {
+                List<Product> products = productRepository.findAll();
+
+                if (products.size() == 0) {
+                    return ResponseEntity.noContent().build();
+                }
+                return ResponseEntity.ok().body(products);
+            }
+        }
         if (pageNum != null) {
             Page<Product> page = productRepository.findAll(PageRequest.of(pageNum.intValue(), 20));
             //
@@ -113,8 +126,16 @@ public class ProductController {
             throw new DuplicateRecordException("Product has already exists");
         }
         Product product = ConvertObject.fromProductDTOToProductDAO(productDTO);
-        Category category = categoryRepository.findById(productDTO.getCategoryId()).get();
-        product.setCategory(category);
+
+        Optional<Category> optionalCategory = categoryRepository.findById(productDTO.getCategoryId());
+
+        if (!optionalCategory.isPresent()) {
+            throw new NotFoundException("Not found category with category id " + productDTO.getCategoryId());
+        }
+
+        product.setCategory(optionalCategory.get());
+
+        CheckValid.checkProduct(product);
 
         Product newProduct = productRepository.save(product);
 
@@ -138,7 +159,11 @@ public class ProductController {
         }
         if (productDTO.getLongDescription() != null) {
             product.setLongDescription(productDTO.getLongDescription().trim().replaceAll("\\s+", " "));
-            product.setShortDescription(product.getLongDescription().substring(0, 60));
+            if (product.getLongDescription().length() > 60) {
+                product.setShortDescription(product.getLongDescription().substring(0, 60));
+            } else {
+                product.setShortDescription(product.getLongDescription());
+            }
         }
         if (productDTO.getCategoryId() != null) {
             Category category = categoryRepository.findById(productDTO.getCategoryId()).get();
@@ -166,6 +191,11 @@ public class ProductController {
         if (!optionalProduct.isPresent()) {
             throw new NotFoundException("Product not found");
         }
+
+        if (!optionalProduct.get().getProductImages().isEmpty()) {
+            throw new InvalidException("Delete failed");
+        }
+
         productRepository.deleteById(productId);
 
         return ResponseEntity.status(HttpStatus.OK).body(optionalProduct.get());
